Defining new contribution process
Android Starlark rules will soon be open sourced and released for customers to use and expand. It is important that we ensure that our user base is well represented and appropriately considered on the path to any contribution. This document serves as a guide for external parties to contribute to Android Starlark Rules productively and efficiently in partnership with the Mobile DevX Build team.