Enable Automatic Exec Groups in rules_android
This is a step forward for the full migration of Automatic Exec Groups.

In this CL, I've enabled AEGs manually (by adding `_use_auto_exec_groups` attribute on rules) for android rules.
I've also added a toolchain_type to `toolchain` attribute of ctx.actions.{run, run_shell} since the Execution platform is selected on a toolchain type level, not on a rule level like before AEGs.

PiperOrigin-RevId: 554457538
Change-Id: Iab178cc170398001d0f20b1593846dcf9f590b06