Run tests with recent java runtime.
The SandboxedSdkToolbox will have dependencies that require more recent
versions of the java runtime to run (>=17), so this change increases the
java runtime version for tests in CI to prepare for that.

Sandboxed SDK rules will also have to use a different java runtime when
running, but this should't affect other rules.

PiperOrigin-RevId: 558063516
Change-Id: Id275701521d7e45856690e3149b4f652bc3c9396