Extract SDK API descriptors from SDK classpath.
This change introduces a new command to the SandboxedSdkToolbox that invokes
the Privacy Sandbox API Packager. The packager is a tool that inspects an SDK's
classpath and extracts tool metadata and classes that are annotated with
Privacy Sandbox annotations.

Note that the packager requires Java runtime version 17 or above to run, so
this change also increases the runtime version in CI.
When running this tool later on we will have to change the runtime version
as well, but this should only impact Sandboxed SDK rules.

PiperOrigin-RevId: 558074024
Change-Id: If147b35fb92699cdfcf97b47ed886e873d3250e0