Increase R8 allowed max memory usage
We've seen this action OOM a few times in our codebase. The default `4g` max set by `java.bzl` probably isn't going to be enough for R8 to successfully dex, desugar, and optimize a large application.

Increasing the max allowed memory should be a safe change since R8 is one of the final actions in an apps build graph.