Always pass proguard implicit output artifacts to Native and symlink them.
When determining implicit outputs, it's not possible to know the configuration of the arrtibute `proguard_specs`, so these implicit output artifacts should always be generated and passed to Native. If the proguard specs turns out to be an empty list at analysis time, they will not be used in dexing and further actions. If users try to request any of them, the build will fail.

PiperOrigin-RevId: 571191172
Change-Id: I14a4200e0e022909c874603c1b7c14187c22a0fe