Add command for generating the RuntimeEnabledSdkConfig message.
This is used in an app bundle to reference the sandboxed SDK dependencies and
how they are signed. This is later read by app stores to properly install
SDK dependencies on devices (or create split APKs with SDK dependencies).

Note that the config message needs to have a `resource id` for every SDK, which
represents the first byte of the package IDs for the SDK. Low values for this
ID can cause problems on API <26 (P and below) so we need to start from a base
value (0x7F) and decrease from there, enforcing a low limit
(50 SDKs). On Android O and above we can use the full byte starting from the
base (0x7F).

Since this logic depends on the minSdkVersion, the command receives the AAPT
manifest dump of the app.

PiperOrigin-RevId: 573756689
Change-Id: I4830042f8356f9a774822f2efedf9fc3011f8221