Use Starlark android_sdk_repository in the android rules
The dexdump file formerly provided by the android sdk supplemental repository is now included in the starlark android_sdk_repository (similar to its native version), and consumers are redirected.
The supplemental repository rule is deleted.
The "files" filegroup was previously unused and working on this CL revealed that build tools post version 30 don't include dx and main dex rules, so they are conditionally removed from the repository as needed

PiperOrigin-RevId: 580256187
Change-Id: Ibc7c5a3d7f473957fab468f582b6de29f7c5b91e