Fix flag extraction bug in extract_desugar_pgcfg_flags
Previous implementation would just blindly split the shrinker_config field as a string by spaces. This logic is incorrect. The right thing to do is to tell the program what format the data is coming in as (i.e. define the struct for the shrinker config flags), and unmarshall the JSON bytes with that in mind.

This CL also enhances the test for this tool to better reflect the actual intended logic.

Part of #122

PiperOrigin-RevId: 601239403
Change-Id: Id4d3e5b7daa92f3d07d588894234f7a2add41f1e