Switch mobile-install Dex Merging to use the dexmerger tool.
This removes the merge archives actions that depends on AK's repack command. Instead we pass all of the transitive dex archives for a shard directly into the DexMerger binary.

In transforms.bzl merge_dex_shards I renamed the variable we pass to merge_dex_shards to make it clear we're passing DEX archive zips to the merging action.

PiperOrigin-RevId: 604682941
Change-Id: I9e7f3c315629db7cb29be25698bdda0ecac53a46