Towards building the example app with bzlmod
The build still relies on a lot of things in WORKSPACE.bzlmod, but this is one step closer. We need all of the stuff in WORKSPACE.bzlmod until the bind()s are removed from android_sdk_repository.

Currently `build` and `mobile-install` still don't work due to a weird issue with python toolchain selection not picking up the custom python toolchain with lxml installed. This is likely due to the aforementioned hacky WORKSPACE.bzlmod integration.

Fixes #206.

PiperOrigin-RevId: 609425942
Change-Id: If6d5d0b8b9f3a8a9e500fcf514ce4dac59ca315a