Use the Android Platforms Transition to add the minSdkVersion to the build configuration.
Add an implicit attribute to read this value in the Dex/Desugar aspect and the Mobile Install aspect. For now we do nothing with the value. The plan is to use the value in select actions in followup changes.

While updating the android_platforms_transition tests I realized I wrote the assertions backwards. So I fixed them for all test cases.

PiperOrigin-RevId: 610820301
Change-Id: I2a54f902bfbba7cc4fee8fa6ba0460f4aa1e194c