Apply the android_transition to many attributes.
This ensures that targets that can be transitively dependend on via deps and any of these attributes are built in at least one matching configuration. I will eventually apply the transition to the manifest attribute once blockers are resolved.

The attributes in attrs.bzl are shared by a few rules. We only apply the transition in the case where they are used by android_binary/android_binary_internal

PiperOrigin-RevId: 623218246
Change-Id: I5764b67c8aed94d8f6939e67f9f43b6ef48c6a39