Get rid of the toplevel starlark android_binary rule.
This toplevel rule was created as a workaround for attempting to apply multiple incoming transitions to the `android_binary` rule: the toplevel `android_binary` applied one and the underlying `android_binary_internal` applied the other. We later found that this caused a configuration issue that the deps of the toplevel rule are configured differently from the underlying `android_binary_internal` rule. As a result, we decide to remove this top-level `android_binary` rule and apply one transition on the incoming edge of `android_binary_internal` rule and the other transition on all the outgoing edges of the rule.

PiperOrigin-RevId: 627493501
Change-Id: I552219aec08307aa1bf916c19ee8216e6036695c