Fix android_binary to access `deps` consistently via `ctx.split_attr`.
Since `deps` has a split transition, `ctx.attr.deps` will return the first branch sorted by their configuration mnemonics. The mnemonic is determined by a number of configuration options, but with the default options, the only difference came down to a hash at the end of the mnemonic, so the choice was basically arbitrary. So when parts of the code that were using a consistent branch from `split_attr` tried to match up artifacts from a different branch from `attr`, the comparisons failed, and misleading errors were thrown. This meant that `--android_platforms` with multiple platforms might arbitrarily fail.

Fixes #230.

PiperOrigin-RevId: 643102707
Change-Id: Ia386407406af1e117866c583e53c2b2abb5f2e65