Change `android_binary` to build against resource classes with non-final fields.
This changes `android_binary`'s behavior to match `android_library`'s in this
respect. After the `android_binary`'s `src` files are compiled, the resource
classes placed into the deploy jar will still have the final fields instead of
non-final.

This provides 2 main benefits:
* Improved consistency between what code can build in an `android_library` and
  `android_binary` target. Currently, the following code can only compile when
  included in `srcs` on an `android_binary` target. It will not compile in an
  `android_library`:
  ```
  swtich (myInt) {
    case R.id.foo:
      ...
  ```
* Improved ability for optimizers to shrink resources. `javac` will inline all
  usages of `static final` fields. This makes it very difficult for an
  optimizer to be sure it accurately identified all usages of resource IDs. If
  the IDs are non-final during `javac`'s execution though, all references to
  the IDs stay as static field references instead of having the constant
  inlined. Optimizers like R8 and ProGuard can optimize away that field read
  anyway so this shouldn't hurt any performance for apps running optimizers.

This change is breaking for any apps that include code like the previous
snippet in an `android_binary` target since that will no-longer compile. The
ACL included in this change should allow opting-out of that behavior to allow
apps to gradually migrate.

PiperOrigin-RevId: 644538754
Change-Id: Ic76cd090d887fb8c16f1095272ea2ec205c20224