Remove an empty line between license comment and docstring
This makes Starlark tools understand docstring, which needs to be on the first uncommented line in the file.
After that load statements follow.

Without this change automatic loads would be inserted before the "docstring", because it's not parsed that way.

The whole change is done automatically using search and replace, without manual intervention.
`s/# limitations under the License.\n\n"/# limitations under the License.\n"/`

PiperOrigin-RevId: 649198497
Change-Id: Ib9a8d964494b2f545ad1346ed56a4256173a795b