Pull aidl_lib from rule attribute rather than the SDK.
As an added benefit we can also delete the android_sdk mobile-install adapter, as this was only used to get the aidl library.

This library is only needed when using the internal compiler (which we plan to delete in favor of the AOSP compiler) so also define an empty library that can be referenced in the OSS rules. This is relatively temporary.

PiperOrigin-RevId: 670753081
Change-Id: I5edbd3eb66414f123f7961a8d311c087aee331b2