Support for desugaring Java 17 Records.
There are three dex flows:

* mobile-install: This uses the dex archive format of the globals
* optimized: this uses the class JAR formate of the globals
* non-optimized: this passes the plain globals directly to the dexmerger

One input to the command to generate the globals file is the min SDK version. For now we have a global value. This means we can define the targets using genrules. If the min SDK version becomes dynamic based on a configuration value the targets will have to be ported to custom rules.

The open source flow doesn't work currently but most of the components are there.

PiperOrigin-RevId: 688746011
Change-Id: I9abfdb209580598f02fd03b884301b3d2f1e7c05