Merge the desugar globals classes into the final APK in a standalone dex shard.
This treats it similarly to the Java 8 legacy dex. Adding the globals during dexmerging was incorrect. It attempted to merge the globals into each shard. The shards are carefully crafted by the DexFileSplitter to avoid going over dex method limits. This led to build failures where we exceeded the limits. It also meant we were merging the globals into every shard instead of just once per APK.

A nice side effect is that this removes the dependency on Bazel 7.4.0+ since we do not need the additional parameter in the android_common.create_dex_merger_actions call.

PiperOrigin-RevId: 691179644
Change-Id: Idd994c1a59286febbfd4398d1e1cd5721198793d